CPM-331: Column required for completeness
